+++ src/types/Variant.ts
@@ -1,5 +1,6 @@
 export const Variant = {
   OUTLINED: 'outlined',
-  PRIMARY: 'primary'
+  PRIMARY: 'primary',
+  TEXT: 'text'
 } as const;
 export type Variant = (typeof Variant)[keyof typeof Variant];