+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Add `Skeleton` component
 
+- Add `IconButton` text variant
+- Add placeholder props for `KeyValueInput` fields
+
 ## v2.0.72
 
 - Add `Panel` component